feat: #WB2-1765, export and add all treeview methods into a treeview util
