+++ packages/react/ui/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from "./MockedDataProvider";
 export * from "./StringUtils";
-export * from "./TreeHelper";
 export * from "./addTimestampToUrl";
 export * from "./blob";
 export * from "./checkUserRight";
@@ -11,4 +10,5 @@ export * from "./noop";
 export * from "./ref";
 export * from "./thumbnail";
 export * from "./time";
+export * from "./treeview";
 export * from "./video";
